@@ -29,10 +29,14 @@ module.exports = function createClient (port_arg, host_arg, options) {
         // [redis:]//[[user][:password]@][host][:port][/db-number][?db=db-number[&password=bar[&option=value]]]
         if (parsed.slashes) { // We require slashes
             if (parsed.auth) {
-                options.password = parsed.auth.split(':')[1];
+                options.password = parsed.auth.slice(parsed.auth.indexOf(':') + 1);
             }
-            if (parsed.protocol && parsed.protocol !== 'redis:') {
-                console.warn('node_redis: WARNING: You passed "' + parsed.protocol.substring(0, parsed.protocol.length - 1) + '" as protocol instead of the "redis" protocol!');
+            if (parsed.protocol) {
+                if (parsed.protocol === 'rediss:') {
+                    options.tls = options.tls || {};
+                } else if (parsed.protocol !== 'redis:') {
+                    console.warn('node_redis: WARNING: You passed "' + parsed.protocol.substring(0, parsed.protocol.length - 1) + '" as protocol instead of the "redis" protocol!');
+                }
             }
             if (parsed.pathname && parsed.pathname !== '/') {
                 options.db = parsed.pathname.substr(1);
@@ -68,7 +72,7 @@ module.exports = function createClient (port_arg, host_arg, options) {
         options.host = options.host || host_arg;
 
         if (port_arg && arguments.length !== 1) {
-            throw new TypeError('To many arguments passed to createClient. Please only pass the options object');
+            throw new TypeError('Too many arguments passed to createClient. Please only pass the options object');
         }
     }
 
