@@ -2,14 +2,13 @@
 
 var util = require('util');
 var assert = require('assert');
-var RedisError = require('redis-parser').RedisError;
+var RedisError = require('redis-errors').RedisError;
 var ADD_STACKTRACE = false;
 
 function AbortError (obj, stack) {
     assert(obj, 'The options argument is required');
     assert.strictEqual(typeof obj, 'object', 'The options argument has to be of type object');
 
-    RedisError.call(this, obj.message, ADD_STACKTRACE);
     Object.defineProperty(this, 'message', {
         value: obj.message || '',
         configurable: true,
