@@ -4,7 +4,9 @@ var index = require('../');
 
 function debug () {
     if (index.debug_mode) {
-        console.error.apply(null, arguments);
+        var data = Array.prototype.slice.call(arguments);
+        data.unshift(new Date().toISOString());
+        console.error.apply(null, data);
     }
 }
 
