@@ -95,7 +95,7 @@ RedisClient.prototype.duplicate = function (options, callback) {
         existing_options[elem] = options[elem];
     }
     var client = new RedisClient(existing_options);
-    client.selected_db = this.selected_db;
+    client.selected_db = options.db || this.selected_db;
     if (typeof callback === 'function') {
         var ready_listener = function () {
             callback(null, client);
