@@ -1,6 +1,6 @@
 'use strict';
 
-var Queue = require('double-ended-queue');
+var Queue = require('denque');
 var utils = require('./utils');
 var Command = require('./command');
 
