@@ -7,10 +7,10 @@ var redis = config.redis;
 
 describe("The 'client' method", function () {
 
-    helper.allTests(function (parser, ip, args) {
+    helper.allTests(function (ip, args) {
         var pattern = /addr=/;
 
-        describe('using ' + parser + ' and ' + ip, function () {
+        describe('using ' + ip, function () {
             var client;
 
             beforeEach(function (done) {
