@@ -8,9 +8,9 @@ var redis = config.redis;
 
 describe("The 'eval' method", function () {
 
-    helper.allTests(function (parser, ip, args) {
+    helper.allTests(function (ip, args) {
 
-        describe('using ' + parser + ' and ' + ip, function () {
+        describe('using ' + ip, function () {
             var client;
             var source = "return redis.call('set', 'sha', 'test')";
 
