@@ -7,9 +7,9 @@ var assert = require('assert');
 
 describe("The 'rpush' command", function () {
 
-    helper.allTests(function (parser, ip, args) {
+    helper.allTests(function (ip, args) {
 
-        describe('using ' + parser + ' and ' + ip, function () {
+        describe('using ' + ip, function () {
             var client;
 
             beforeEach(function (done) {
