@@ -7,9 +7,9 @@ var redis = config.redis;
 
 describe("The 'zadd' method", function () {
 
-    helper.allTests(function (parser, ip, args) {
+    helper.allTests(function (ip, args) {
 
-        describe('using ' + parser + ' and ' + ip, function () {
+        describe('using ' + ip, function () {
             var client;
 
             beforeEach(function (done) {
@@ -34,7 +34,11 @@ describe("The 'zadd' method", function () {
                 client.zrange('infinity', 0, -1, 'WITHSCORES', function (err, res) {
                     assert.equal(res[5], 'inf');
                     assert.equal(res[1], '-inf');
-                    assert.equal(res[3], '9.9999999999999992e+22');
+                    if (process.platform !== 'win32') {
+                        assert.equal(res[3], '9.9999999999999992e+22');
+                    } else {
+                        assert.equal(res[3], '9.9999999999999992e+022');
+                    }
                     done();
                 });
             });
