@@ -161,13 +161,7 @@ module.exports = {
             cb = opts;
             opts = {};
         }
-        var parsers = ['javascript'];
         var protocols = ['IPv4'];
-        // The js parser works the same as the hiredis parser, just activate this if you want to be on the safe side
-        // try {
-        //     require('hiredis');
-        //     parsers.push('hiredis');
-        // } catch (e) {/* ignore eslint */}
         if (process.platform !== 'win32') {
             protocols.push('IPv6', '/tmp/redis.sock');
         }
@@ -185,13 +179,11 @@ module.exports = {
                 }
             }
             describe('using options: ' + strOptions, function () {
-                parsers.forEach(function (parser) {
-                    protocols.forEach(function (ip, i) {
-                        if (i !== 0 && !opts.allConnections) {
-                            return;
-                        }
-                        cb(parser, ip, config.configureClient(parser, ip, options));
-                    });
+                protocols.forEach(function (ip, i) {
+                    if (i !== 0 && !opts.allConnections) {
+                        return;
+                    }
+                    cb(ip, config.configureClient(ip, options));
                 });
             });
         });
