@@ -7,9 +7,9 @@ var redis = config.redis;
 
 describe('publish/subscribe', function () {
 
-    helper.allTests(function (parser, ip, args) {
+    helper.allTests(function (ip, args) {
 
-        describe('using ' + parser + ' and ' + ip, function () {
+        describe('using ' + ip, function () {
             var pub = null;
             var sub = null;
             var channel = 'test channel';
@@ -515,8 +515,8 @@ describe('publish/subscribe', function () {
                         sub.stream.once('data', function () {
                             assert.strictEqual(sub.message_buffers, false);
                             assert.strictEqual(sub.shouldBuffer, false);
-                            sub.on('pmessageBuffer', function (pattern, channel, message) {
-                                if (parser !== 'javascript' && typeof pattern === 'string') {
+                            sub.on('pmessageBuffer', function (pattern, channel) {
+                                if (typeof pattern === 'string') {
                                     pattern = new Buffer(pattern);
                                     channel = new Buffer(channel);
                                 }
@@ -524,10 +524,8 @@ describe('publish/subscribe', function () {
                                 assert.strictEqual(channel.inspect(), new Buffer('/foo').inspect());
                                 sub.quit(end);
                             });
-                            if (parser === 'javascript') {
-                                assert.notStrictEqual(sub.message_buffers, sub.buffers);
-                            }
-
+                            // Either message_buffers or buffers has to be true, but not both at the same time
+                            assert.notStrictEqual(sub.message_buffers, sub.buffers);
                         });
                         var batch = sub.batch();
                         batch.psubscribe('*');
