@@ -7,6 +7,7 @@ var helper = require('./helper');
 var path = require('path');
 var redis = config.redis;
 var utils = require('../lib/utils');
+var tls = require('tls');
 
 var tls_options = {
     servername: 'redis.js.org',
@@ -18,9 +19,6 @@ var tls_port = 6380;
 // Use skip instead of returning to indicate what tests really got skipped
 var skip = false;
 
-// Wait until stunnel4 is in the travis whitelist
-// Check: https://github.com/travis-ci/apt-package-whitelist/issues/403
-// If this is merged, remove the travis env checks
 describe('TLS connection tests', function () {
 
     before(function (done) {
@@ -28,9 +26,6 @@ describe('TLS connection tests', function () {
         if (process.platform === 'win32') {
             skip = true;
             console.warn('\nStunnel tests do not work on windows atm. If you think you can fix that, it would be warmly welcome.\n');
-        } else if (process.env.TRAVIS === 'true') {
-            skip = true;
-            console.warn('\nTravis does not support stunnel right now. Skipping tests.\nCheck: https://github.com/travis-ci/apt-package-whitelist/issues/403\n');
         }
         if (skip) return done();
         helper.stopStunnel(function () {
@@ -57,7 +52,7 @@ describe('TLS connection tests', function () {
             client = redis.createClient({
                 connect_timeout: connect_timeout,
                 port: tls_port,
-                tls: tls_options
+                tls: utils.clone(tls_options)
             });
             var time = 0;
             assert.strictEqual(client.address, '127.0.0.1:' + tls_port);
@@ -90,12 +85,12 @@ describe('TLS connection tests', function () {
 
         it('connect with host and port provided in the tls object', function (done) {
             if (skip) this.skip();
-            var tls = utils.clone(tls_options);
-            tls.port = tls_port;
-            tls.host = 'localhost';
+            var tls_opts = utils.clone(tls_options);
+            tls_opts.port = tls_port;
+            tls_opts.host = 'localhost';
             client = redis.createClient({
                 connect_timeout: 1000,
-                tls: tls
+                tls: tls_opts
             });
 
             // verify connection is using TCP, not UNIX socket
@@ -106,6 +101,30 @@ describe('TLS connection tests', function () {
 
             client.set('foo', 'bar');
             client.get('foo', helper.isString('bar', done));
+        });
+
+        describe('using rediss as url protocol', function () {
+            var tls_connect = tls.connect;
+            beforeEach(function () {
+                tls.connect = function (options) {
+                    options = utils.clone(options);
+                    options.ca = tls_options.ca;
+                    options.servername = 'redis.js.org';
+                    options.rejectUnauthorized = true;
+                    return tls_connect.call(tls, options);
+                };
+            });
+            afterEach(function () {
+                tls.connect = tls_connect;
+            });
+            it('connect with tls when rediss is used as the protocol', function (done) {
+                if (skip) this.skip();
+                client = redis.createClient('rediss://localhost:' + tls_port);
+                // verify connection is using TCP, not UNIX socket
+                assert(client.stream.encrypted);
+                client.set('foo', 'bar');
+                client.get('foo', helper.isString('bar', done));
+            });
         });
 
         it('fails to connect because the cert is not correct', function (done) {
