@@ -1,7 +1,7 @@
 'use strict';
 
 var assert = require('assert');
-var Queue = require('double-ended-queue');
+var Queue = require('denque');
 var utils = require('../lib/utils');
 var intercept = require('intercept-stdout');
 
@@ -131,17 +131,22 @@ describe('utils.js', function () {
         });
 
         it('elements in the offline queue. Reply after the offline queue is empty and respect the command_obj callback', function (done) {
-            clientMock.offline_queue.push(create_command_obj(), create_command_obj());
+            clientMock.offline_queue.push(create_command_obj());
+            clientMock.offline_queue.push(create_command_obj());
             utils.reply_in_order(clientMock, function () {
                 assert.strictEqual(clientMock.offline_queue.length, 0);
                 assert.strictEqual(res_count, 2);
                 done();
             }, null, null);
-            while (clientMock.offline_queue.length) clientMock.offline_queue.shift().callback(null, 'foo');
+            while (clientMock.offline_queue.length) {
+                clientMock.offline_queue.shift().callback(null, 'foo');
+            }
         });
 
         it('elements in the offline queue. Reply after the offline queue is empty and respect the command_obj error emit', function (done) {
-            clientMock.command_queue.push({}, create_command_obj(), {});
+            clientMock.command_queue.push({});
+            clientMock.command_queue.push(create_command_obj());
+            clientMock.command_queue.push({});
             utils.reply_in_order(clientMock, function () {
                 assert.strictEqual(clientMock.command_queue.length, 0);
                 assert(emitted);
@@ -158,8 +163,10 @@ describe('utils.js', function () {
         });
 
         it('elements in the offline queue and the command_queue. Reply all other commands got handled respect the command_obj', function (done) {
-            clientMock.command_queue.push(create_command_obj(), create_command_obj());
-            clientMock.offline_queue.push(create_command_obj(), {});
+            clientMock.command_queue.push(create_command_obj());
+            clientMock.command_queue.push(create_command_obj());
+            clientMock.command_queue.push(create_command_obj());
+            clientMock.offline_queue.push({});
             utils.reply_in_order(clientMock, function (err, res) {
                 assert.strictEqual(clientMock.command_queue.length, 0);
                 assert.strictEqual(clientMock.offline_queue.length, 0);
